Refactor google coverart javascript to use Stimulus
