+++ app/javascript/controllers/index.js
@@ -31,6 +31,9 @@ application.register("course-reserves", CourseReservesController)
 import GalleryPreviewController from "./gallery_preview_controller"
 application.register("gallery-preview", GalleryPreviewController)
 
+import GoogleCoverImageController from "./google_cover_image_controller"
+application.register("google-cover-image", GoogleCoverImageController)
+
 import LiveLookupController from "./live_lookup_controller"
 application.register("live-lookup", LiveLookupController)
