+++ app/views/catalog/_document_gallery.html.erb
@@ -3,23 +3,25 @@
 
 <% # container for all documents in index view -%>
 <% if params[:controller] == "browse" && params[:action] == "nearby" %>
-  <div class="gallery-document">
-    <div class="browse-link">
-      <%= link_to 'Continue to full page', browse_index_path(
-        start: params[:start],
-        call_number: params[:call_number],
-        view: :gallery
-      ), class: 'text-center' %>
+  <div data-controller="google-cover-image" class="d-flex">
+    <div class="gallery-document">
+      <div class="browse-link">
+        <%= link_to 'Continue to full page', browse_index_path(
+          start: params[:start],
+          call_number: params[:call_number],
+          view: :gallery
+        ), class: 'text-center' %>
+      </div>
     </div>
-  </div>
-  <%= render view_config.document_component.with_collection(document_presenters, partials: view_config.partials, counter_offset: @response&.start || 0) %>
-  <div class="gallery-document">
-    <div class="browse-link">
-      <%= link_to 'Continue to full page', browse_index_path(
-        start: params[:start],
-        call_number: params[:call_number],
-        view: :gallery
-      ), class: 'text-center' %>
+    <%= render view_config.document_component.with_collection(document_presenters, partials: view_config.partials, counter_offset: @response&.start || 0) %>
+    <div class="gallery-document">
+      <div class="browse-link">
+        <%= link_to 'Continue to full page', browse_index_path(
+          start: params[:start],
+          call_number: params[:call_number],
+          view: :gallery
+        ), class: 'text-center' %>
+      </div>
     </div>
   </div>
 <% elsif %>
