+++ app/views/layouts/searchworks.html.erb
@@ -65,7 +65,7 @@
           <%= render 'shared/search_subnavbar' %>
 
           <%= render partial: 'shared/modal' %>
-          <main id="main-container">
+          <main id="main-container" data-controller="google-cover-image">
             <%= render 'global_alerts/alerts' if Settings.GLOBAL_ALERT %>
             <%= render :partial=>'/flash_msg', layout: 'shared/flash_messages' %>
